-corrigido cabeçalho do instalador -verifica se o wget ou curl estão instalados -verifica se o mmaker esta instalado -se o instalador do ohmyzsh não for obtido,ignorar instalação -usar wget e curl no modo silencioso -checar se o wget e curl estão instalados de modo silencioso -adicionado função de backup

+++ install.sh
@@ -1,14 +1,51 @@
 #!/usr/bin/env bash
-:'
+: '
 dotfiles installer
+
 '
+#======= constants =======
+readonly OHMYZSH_DOWNLOAD_LINK="https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh"
+#======= variables =======
+#======= flags =======
+flag_ok_ohmyzsh=0
+flag_make_copy=0 #copy the config files
+flag_make_xinit=0 #create the .xinitrc file
+flag_make_menu=0 #create a new menu file
 #======= functions ======= 
+function backup(){
+	#create a backup of current config files
+	if [ -e  $HOME/.zshrc ];then
+		cp -v $HOME/.zshrc $HOME/.zshrc.bak
+	fi
+	if [ -e  $HOME/.bashrc ];then
+		cp -v $HOME/.bashrc $HOME/.bashrc.bak
+	fi
+	if [ -e  $HOME/.vimrc ];then
+		cp -v $HOME/.vimrc $HOME/.vimrc.bak
+	fi
+	if [ -e  $HOME/.tmux.conf ];then
+		cp -v $HOME/.tmux.conf $HOME/.tmux.conf.bak
+	fi
+	if [ -e  $HOME/.xinitrc ];then
+		cp -v $HOME/.xinitrc $HOME/.xinitrc.bak
+	fi
+}
+
 function copy(){
 	#install ohmyzsh
 	echo "download ohmyzsh..."
-	wget -O $HOME/install_ohmyzsh.sh https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
-	chmod +x $HOME/install_ohmyzsh.sh 
-	cp -v $HOME/.zshrc $HOME/.zshrc.bak
+	if command -v wget 1>>/dev/null 2>>/dev/null ;then
+		#wget -O $HOME/install_ohmyzsh.sh https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
+		eval wget -q -O $HOME/install_ohmyzsh.sh "$OHMYZSH_DOWNLOAD_LINK"
+		chmod +x $HOME/install_ohmyzsh.sh 
+	elif command -v curl 1>>/dev/null 2>>/dev/null ;then
+		eval curl -S --show-error $OHMYZSH_DOWNLOAD_LINK -o $HOME/install_ohmyzsh.sh
+		chmod +x $HOME/install_ohmyzsh.sh 
+	else
+		echo "wget and curl is not installed"
+		echo "cannot get OhMyZsh, ignoring..."
+	fi
+
 	cp -v vimrc $HOME/.vimrc
 	cp -v tmux.conf $HOME/.tmux.conf
 	cp -v zshrc $HOME/.zshrc
@@ -25,20 +62,29 @@ function copy(){
 	cp -v bashrc $HOME/.bashrc
 	cp -v xprofile $HOME/.xprofile
 	cp -v -r cool-retro-term $HOME/.local/share/cool-retro-term
-	echo "install ohmyzsh"
-	$HOME/install_ohmyzsh.sh --unattended --keep-zshrc
+	if [ $flag_ok_ohmyzsh -eq 0 ];then
+		echo "install ohmyzsh"
+		$HOME/install_ohmyzsh.sh --unattended --keep-zshrc
+	else
+		echo "not installing ohmyzsh"
+	fi
 }
 
 function mk_xinit(){
-	echo "openbox-session" >> $HOME/.xinitrc
+	echo "openbox-session" > $HOME/.xinitrc
 }
 function create_menu(){
-	eval  mmaker -vf OpenBox3
+	if command -v mmaker ;then
+		eval  mmaker -vf OpenBox3
+	else
+		echo "error: menumaker(mmaker) is not installed"
+	fi
 }
 function usage(){
 	echo "dotfiles install script"
 	echo "./install [options]"
 	echo "options:"
+	echo "	-b,--backup : backup of current config files"
 	echo "	-i,--install : install configfiles"
 	echo "	-m,--newmenu : create a new menu file"
 	echo "	-x,--make-init : create .xinitrc file"
@@ -48,6 +94,9 @@ function usage(){
 #======= start execution ======= 
 while [ "$1" != "" ];do
 	case $1 in
+		-b | --backup )
+			backup
+			;;
 		-i | --install )
 			copy
 			;;
@@ -64,6 +113,6 @@ while [ "$1" != "" ];do
 		* )
 			echo "option \"$1\" not exist"
 			exit 1
-	shift
 	esac
+	shift
 done
